Modularize shared note delivery API
